+++ lib/onyxcord/application_commands.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'onyxcord/application_commands/option'
+require 'onyxcord/application_commands/command'
+require 'onyxcord/application_commands/context'
+require 'onyxcord/application_commands/registry'
+
+module OnyxCord
+  module ApplicationCommands
+  end
+end