+++ lib/onyxcord/websocket.rb
@@ -1,25 +1,15 @@
 # frozen_string_literal: true
 
-require 'async'
+require 'onyxcord/async/runtime'
 require 'async/http/endpoint'
 require 'async/websocket/client'
 
 module OnyxCord
-  # Wrapper around async-websocket that provides the same callback-based
-  # interface used by the Gateway and Voice subsystems. This replaces the
-  # previous websocket-client-simple implementation.
   class WebSocket
-    # @return [Boolean] whether the connection is currently open.
     attr_reader :connected
 
     alias_method :connected?, :connected
 
-    # Creates a new WebSocket wrapper.
-    # @param host [String]            The `wss://` endpoint URL to connect to.
-    # @param open_handler [Proc]      Called once the connection is established.
-    # @param message_handler [Proc]   Called for every text frame received.
-    # @param error_handler [Proc]     Called when an error occurs.
-    # @param close_handler [Proc]     Called when the connection closes.
     def initialize(host, open_handler, message_handler, error_handler, close_handler)
       @host = host
       @open_handler = open_handler
@@ -33,8 +23,6 @@ module OnyxCord
       connect
     end
 
-    # Send a text message over the WebSocket.
-    # @param data [String, Hash] Data to send. Hashes are JSON-encoded automatically.
     def send(data)
       return unless @connection
 
@@ -45,7 +33,6 @@ module OnyxCord
       @error_handler&.call(e)
     end
 
-    # Cleanly close the connection.
     def close
       @connected = false
       @connection&.close
@@ -58,7 +45,7 @@ module OnyxCord
     def connect
       endpoint = Async::HTTP::Endpoint.parse(@host)
 
-      Async do
+      @task = OnyxCord::AsyncRuntime.async do
         Async::WebSocket::Client.connect(endpoint) do |connection|
           @connection = connection
           @connected = true